[varLib cff2] Sparse source support
Fix for bug in models._computeMasterSupports() for fonts with more than one  intermediate master on the axis. The current code resets the "lower" end of the (lower, peak, upper) triplet based on the previous master, but never resets the 'upper' value. (+1 squashed commit)
Squashed commits:
[23e90064] [varLib cff2] Sparse source support

Fix for bug in models._computeMasterSupports() for fonts with more than one intermediate master on the axis. The current code resets the "lower" end of the (lower, peak, upper) triplet based on the previous master, but never resets the 'upper' value.