[varLib cff2] Sparse source support
When converting between T2Charstring 'program' and 'commands' lists with specializer.py programToCommands() and commandsToProgram(), map each blend operator and its arguments in 'program' to a blend argument list in 'commands'. This allows specializeCommands() and generalizeCommands() to work with charstrings that have blend operators.

Update specializeCommands() to combine two successive [vh]line operators when one or both of the two arguments is a blend argument list.

This uses a new class in varStore.py to hold the delta weights for a specific vsIndex.